% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_overlap.R
\name{circular_overlap_24hour}
\alias{circular_overlap_24hour}
\title{Overlap of Two Hourly Circular Distributions}
\usage{
circular_overlap_24hour(a, b, normal = TRUE)
}
\arguments{
\item{a}{A vector whose length equals the number of individuals within a community with entries containing
one measurement per individual of a certain trait.}

\item{b}{Another vector whose length equals the number of individuals within a community with entries containing the same trait measurements per individual to be compared against a.}

\item{normal}{If set to TRUE, then the area under all density functions is normalized to one. If set to FALSE,
then the area under all density functions is proportional to the number of observations in that group.}
}
\value{
The function returns a vector of three values:
\item{overlap_average}{The average overlap of a and b, calculated by the overlap
area *2 divided by the sum of the areas under the two functions.}
\item{overlap_a}{The proportion of a that overlaps with b, calculated by the overlap
area divided by the area under the function generated from a.}
\item{overlap_b}{The proportion of b that overlaps with a, calculated by the overlap
area divided by the area under the function generated from b.}
}
\description{
This function calculates circular overlap based on discrete hourly observations from hours
0 to 23 (24 hours per day).
}
\details{
This function analyzes discrete data collected on the hour. Data are converted
to a factor with levels(0:23). The function then by manually calculates the density by taking
the proportion of each hour. The intersection density function is then calculated by
taking the integral of the minimum of the two functions, from which the overlap
outputs are calculated.
}
\examples{
# Load dplyr library to subset ant_data.
library(dplyr)

# Vector a will represent all ant_data from experimental chamber 1.
a <- filter(ant_data, chamber == 1)

# Vector b will represent all ant_data from experimental chamber 2.
b <- filter(ant_data, chamber == 2)

# Calculate the circular overlap between vectors a and b.
circular_overlap_24hour(a$time,b$time)

}
\seealso{
\code{\link{pairwise_overlap}} To calculate the linear overlap between two empirical
density estimates.

\code{\link{circular_overlap}} To calculate continous circular overlap between
two empirical density estimates.
}
