% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityoverlap.R
\name{community_overlap}
\alias{community_overlap}
\title{Median pairwise overlap across all Species in a community}
\usage{
community_overlap(traits, sp, normal = TRUE, bw = NULL, N = NULL)
}
\arguments{
\item{traits}{A vector whose length equals the number of individuals in a community with entries of one trait measurement for each individual.}

\item{sp}{A vector with length equal to length(traits) indicating the taxon of each individual.}

\item{bw}{The smoothing bandwidth to be used by the density function. The kernels are scaled, so
that the smoothing bandwidth is the standard deviation of the smoothing kernel.}

\item{norm}{Set to TRUE when the trait data are normally distributed. If the trait data are not normally distributed, then an additional normalization step is carried out by setting this parameter to FALSE and each density will be multiplied by the length of each vector.}

\item{n}{The number of equally spaced points at which the density is to be
estimated.}
}
\value{
The function returns a median of pairwise overlaps of all species in
the community.
}
\description{
This function calculates the median of pairwise overlaps between density
estimates of trait distributions of all species within a community.
}
\details{
The funtion evaluates pairwise overlaps of density estimates of all species with abundances >1 in a community to calculate the median overlap across pairs for the community.
}
\seealso{
\code{\link{pairwise_overlap}} To calculate overlap between two empirical
density estimates.

\code{\link{community_overlap_harmonicwmedian}} To use harmonic means of abundances of species pairs as weights to calculate the median overlap.
}
