% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_overlap.R
\name{community_overlap_circular}
\alias{community_overlap_circular}
\title{Community-level weighted median overlap using circular distributions}
\usage{
community_overlap_circular(traits, sp, norm = TRUE, randomize_weights = FALSE)
}
\description{
Defaults to use 24-hour clock as units.
Uses manual calculation of density.
}
