% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityoverlap.R
\name{community_overlap_harmonicwmedian}
\alias{community_overlap_harmonicwmedian}
\title{Abundance-weighted (harmonic) median pairwise overlap across all species in a community}
\usage{
community_overlap_harmonicwmedian(
  traits,
  sp,
  norm = TRUE,
  bw = NULL,
  n = NULL,
  randomize_weights = FALSE
)
}
\description{
Input: vector of traits, vector of species identities
}
\details{
This version uses the harmonic mean of abundances in each species pair as weights.
}
