% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityoverlap.R
\name{community_overlap_noitv}
\alias{community_overlap_noitv}
\title{Abundance-weighted median of pairwise distance between species trait means
in a Community}
\usage{
community_overlap_noitv(traits, sp, output = "median", weight_type = "hmean")
}
\arguments{
\item{traits}{A numeric vector of trait measurements with length equal to the number of individuals. Each individual should have a trait measurement.}

\item{sp}{A factor vector with length equal to the number of elements in the traits vector that indicates the taxon of each individual.}

\item{output}{The arguments *output* specifies the whether the median or mean of all pairwise overlap values between distributions of species will be returned. The default value for output is "median".}

\item{weight_type}{Specifies weights to be used to calculate the median or mean output. When the default value for *weight_type* = "hmean", the function calculates the pairwise overlaps of trait distributions of all species in each community as 2/(1/abundance of species a + 1/abundance of species b). The harmonic mean *weight_type* is set as the default as it minimizes the effect of outliers and rare species. If the argument weight_type = "none", no weights are used for the calculation of mean or median. If weight_type = "mean", raw abundances of species are used as weights.}
}
\value{
The function returns a mean or median of pairwise distances of all species in
the community, which can be weighted by harmonic means or arithmetic means of abundances
of species pairs.
}
\description{
This function calculates the median of pairwise distances between species
trait means within a community, weighted by the mean of abundances of the
species pairs. This reduces each species to a trait mean and the intraspecific
trait variation (ITV) is ignored for comparison to an analysis with ITV considered.
}
\details{
The function evaluates the means of trait measurements of each
species in a community with abundance >1 and calculates weighted mean or median of pairwise
distances between all species in a community.
}
\note{
This function so far only works for linear datasets.
}
\examples{
library(tidyverse)
library(Ostats)

# Load data from web archive and Keep only the relevant part of data
dat <- read_csv('https://ndownloader.figshare.com/files/9167548') \%>\%
   filter(siteID \%in\% 'HARV') \%>\%
   select(siteID, taxonID, weight) \%>\%
   filter(!is.na(weight)) \%>\%
   mutate(log_weight = log10(weight))

# Calculate median of pairwise distances for the community,weighted by harmonic means
of abundances
community_overlap_noitv(traits = as.matrix(dat$log_weight),
   sp = factor(dat$taxonID))

}
\seealso{
\code{\link{community_overlap_merged}} To see a function that takes
intraspecific trait variation (ITV) into account to calculate overlap statistics
for a community.
\code{\link{Ostats_plot}} For plotting community overlap for
each community.
}
