% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityoverlap.R
\name{community_overlap_wmedian}
\alias{community_overlap_wmedian}
\title{Abundance-weighted median pairwise overlap across all species in a community}
\usage{
community_overlap_wmedian(traits, sp, norm = TRUE, bw = NULL, n = NULL)
}
\description{
Input: vector of traits, vector of species identities
}
