<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_ses}
\alias{get_ses}
\title{Calculation of standardized effect sizes of overlap statistics against null distributions}
\usage{
get_ses(o, o_n, qs)
}
\arguments{
\item{o}{an matrix data containing the overlapping values of normal distribution of the observed data (the observed local O-stats for each community)}

\item{o_n}{an array object containing the overlapping values of normal distribution of the randomize data - (the null distributions of O-stats function)}

\item{qs}{quantiles of the null distribution to extract, usually 0.025 and 0.975}
}
\value{
ses - a matrix containing values of standardized effect sizes for observed data of each of the locals (rows) and traits(columns) avaliated. To see if the overlap values can be expected by chance see returns ses_upper and ses_lower. If the value corresponding to the local in ses is between ses_upper and ses_lower, the overlapping between traits in community can be expected by chance.

ses_lower
a matrix containing null values extracted for the quantil of qs[1] (usually 0.025)

ses_upper
a matrix containing null values extracted for the quantil of qs[2] (usually 0.975)

raw_lower
a matrix containing null values extracted for the quantil of qs[1] (usually 0.025) without subtracting the mean and divided by standard deviation (see equation for calculating the standardize effect size on the function code)

raw_upper
a matrix containing null values extracted for the quantil of qs[2] (usually 0.975) without subtracting the mean and divided by standard deviation (see equation for calculating the standardize effect size on the function code)
}
\description{
Extract quantiles to get standardized effect sizes for the overlap stats
}
\details{
Implementation of null model that swaps means of distributions to test whether distributions are more evenly spaced than expected by chance - It is part of the Ostats function
}
\seealso{
\code{\link{Ostats}} to Calculate O-statistics (community-level pairwise niche overlap statistics)
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_ses}
\alias{get_ses}
\title{Calculation of standardized effect sizes of overlap statistics against null distributions}
\usage{
get_ses(o, o_n, qs)
}
\arguments{
\item{o}{an matrix data containing the overlapping values of normal distribution of the observed data (the observed local O-stats for each community)}

\item{o_n}{an array object containing the overlapping values of normal distribution of the randomize data - (the null distributions of O-stats function)}

\item{qs}{quantiles of the null distribution to extract, usually 0.025 and 0.975}
}
\value{
ses - a matrix containing values of standardized effect sizes for observed data of each of the locals (rows) and traits(columns) avaliated. To see if the overlap values can be expected by chance see returns ses_upper and ses_lower. If the value corresponding to the local in ses is between ses_upper and ses_lower, the overlapping between traits in community can be expected by chance.

ses_lower
a matrix containing null values extracted for the quantil of qs[1] (usually 0.025)

ses_upper
a matrix containing null values extracted for the quantil of qs[2] (usually 0.975)

raw_lower
a matrix containing null values extracted for the quantil of qs[1] (usually 0.025) without subtracting the mean and divided by standard deviation (see equation for calculating the standardize effect size on the function code)

raw_upper
a matrix containing null values extracted for the quantil of qs[2] (usually 0.975) without subtracting the mean and divided by standard deviation (see equation for calculating the standardize effect size on the function code)
}
\description{
Extract quantiles to get standardized effect sizes for the overlap stats
}
\details{
Implementation of null model that swaps means of distributions to test whether distributions are more evenly spaced than expected by chance - It is part of the Ostats function
}
\seealso{
\code{\link{Ostats}} to Calculate O-statistics (community-level pairwise niche overlap statistics)
}
>>>>>>> f1f75868774584212ae6ad6f9a2cb0d91b8f9641
