% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats2.R
\name{Ostats2}
\alias{Ostats2}
\title{Calculate O-statistics (community-level pairwise niche overlap statistics)}
\usage{
Ostats2(
  traits,
  plots,
  sp,
  overlap_function = pairwise_overlap,
  central_tendency = c("median", "mean"),
  abundance_weighted = TRUE,
  weighting_type = c("harmonic", "arithmetic"),
  nperm = 99,
  nullqs = c(0.025, 0.975),
  shuffle_weights = FALSE,
  swap_means = FALSE
)
}
\description{
This is the primary function in the Ostats package. It evaluates the
O-statistics against a local null model.
}
\details{
This test version created by QDR on 20 April 2020 is more flexible in two ways:
First, you can choose any measure of central tendency for calculating the
community-level overlap statistic from the vector of pairwise overlaps.
Second, you can choose any function for calculating the overlap. This is relevant
if you are doing the overlaps in circular distributions.

#### STILL IN DEVELOPMENT -- NOT DONE!!! -Q 4/20
}
