% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats.R
\name{Ostats_circular}
\alias{Ostats_circular}
\title{Calculate O-statistics using circular distributions}
\usage{
Ostats_circular(
  traits,
  plots,
  sp,
  nperm = 99,
  nullqs = c(0.025, 0.975),
  shuffle_weights = FALSE,
  swap_means = FALSE
)
}
\description{
This is a version of the O-statistics function that works with circular distributions.
Circular distributions are only implemented for the local null model, not regional.
}
\details{
Defaults to use 24-hour clock as units for the circular distribution.
This is a simplified non-flexible version that only does the "manual" density calculation.
}
\seealso{
\code{\link{Ostats}} for non-circular distributions.
}
