% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats_multivariate.R
\name{Ostats_multivariate}
\alias{Ostats_multivariate}
\title{Test version of multivariate O-stats}
\usage{
Ostats_multivariate(
  traits,
  plots,
  sp,
  output = "median",
  weight_type = "hmean",
  nperm = 99,
  nullqs = c(0.025, 0.975),
  shuffle_weights = FALSE,
  swap_means = FALSE,
  hypervolume_args = list()
)
}
\description{
QDR 19 Oct 2020
I decided to make this a separate function for now, but it could potentially be merged into the main Ostats() function later
However, the data structures are somewhat different so it would take a bit of coding to merge them.
This does not support circular data.
Also note at the moment that the argument hypervolume_args is passed to density_args for consistency with argument names.
It is all internal so not really an issue.
}
