% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats_multivariate_plot.R
\name{Ostats_multivariate_plot}
\alias{Ostats_multivariate_plot}
\title{Plot multivariate community overlap}
\usage{
Ostats_multivariate_plot(
  plots,
  sp,
  traits,
  overlap_dat = NULL,
  use_plots = NULL,
  colorvalues = NULL,
  plot_points = TRUE,
  panel_height = 3,
  panel_width = 3,
  units = "cm",
  hypervolume_args = list()
)
}
\arguments{
\item{plots}{Site identity: a vector of names of each community.}

\item{sp}{Taxon identity: a vector of species or taxa names.}

\item{traits}{A matrix or data frame with rows representing individuals and columns representing traits.}

\item{overlap_dat}{Optional: an object containing the output of \code{\link{Ostats_multivariate}}. If provided, overlap statistics will be displayed in the plot panels.}

\item{use_plots}{a vector of sites to plot. If NULL, the function will plot all the sites.}

\item{colorvalues}{Vector of color values for the density polygons. Defaults to a viridis palette if none provided.}

\item{plot_points}{whether to plot individual data points in addition to the hypervolume slices. Default is TRUE.}

\item{panel_height}{height of the individual plot panels, in units given by \code{units}. Default is 3 cm.}

\item{panel_width}{height of the individual plot panels, in units given by \code{units}. Default is 3 cm.}

\item{units}{units for panel height and width. Default is centimeters.}

\item{hypervolume_args}{additional arguments to pass to \code{\link[hypervolume]{hypervolume}},
such as \code{method}. If none are provided, default values
are used.

MORE DOCUMENTATION GOES HERE}
}
\description{
Placeholder.
}
