% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_overlap.R
\name{pairwise_overlap}
\alias{pairwise_overlap}
\title{Overlap between Two Empirical Density Estimates}
\usage{
pairwise_overlap(a, b, density_args = list())
}
\description{
This function generates kernel density estimates for two datasets on a common
grid to calculate the area of overlap between the two estimates.
See \url{http://stats.stackexchange.com/questions/97596/how-to-calculate-overlap-between-empirical-probability-densities},
this function is derived from code posted in the answer by user mmk.
}
\details{
FIXME new documentation could be added here, if we want to still export this function.
}
