% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_pool.R
\docType{data}
\name{reg_pool}
\alias{reg_pool}
\title{Sample Data for Ostats_regional}
\format{
An object of class \code{data.frame} with 7905 rows and 6 columns.
}
\usage{
reg_pool
}
\description{
This is a sample dataset with small mammal data from the NEON sites BART and
HARV across years.The two sites belong to the same domain, the Northeast domain.
}
\references{

}
\keyword{datasets}
