% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_mammal_data.R
\docType{data}
\name{small_mammal_data}
\alias{small_mammal_data}
\title{Sample Continuous Trait Data (NEON Small Mammals 2015)}
\format{
An object of class \code{data.frame} with 7692 rows and 7 columns.
}
\usage{
small_mammal_data
}
\description{
Placeholder.
}
\references{
Placeholder.
}
\keyword{datasets}
