% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_mammal_Ostats.R
\docType{data}
\name{small_mammal_Ostats}
\alias{small_mammal_Ostats}
\title{Pre-calculated Overlap Statistics for NEON Small Mammals 2015}
\format{
An object of class \code{list} of length 4.
}
\usage{
small_mammal_Ostats
}
\description{
Placeholder.
}
\references{
Placeholder.
}
\keyword{datasets}
