% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant_data.R
\docType{data}
\name{ant_data}
\alias{ant_data}
\title{Sample Circular Data}
\format{
An object of class \code{data.frame} with 4837 rows and 3 columns.
}
\usage{
ant_data
}
\description{
This is a sample dataset with discrete hourly data collection from a study
in which ants were observed during 24-hour periods in multiple experimental
chambers. Each row represents an observation of an individual ant, with its
species identity and the time it was observed, rounded to the nearest hour.
The experimental setup is described in Stuble et al. 2014.
}
\references{
Stuble, K.L., Patterson, C.M., Rodriguez-Cabal, M.A., Ribbons,
  R.R., Dunn, R.R. & Sanders, N.J. (2014). Ant-mediated seed dispersal in a
  warmed world. PeerJ, 2, e286. https://peerj.com/articles/286/
}
\keyword{datasets}
