% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_overlap.R
\name{circular_overlap}
\alias{circular_overlap}
\title{Overlap of two circular distributions}
\usage{
circular_overlap(a, b, circular_units, circular_template, norm = TRUE,
  bw, n = NULL)
}
\description{
This function converts the input vectors to circular objects, calculates empirical densities,
and then calculates their overlap.
}
\details{
The user must specify the bandwidth for the KDE as well as some other options for the circular
conversion.
}
