% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_overlap.R
\name{circular_overlap_24hour}
\alias{circular_overlap_24hour}
\title{Overlap of two 24-hour distributions}
\usage{
circular_overlap_24hour(a, b, norm = TRUE)
}
\description{
This manually calculates the density by just taking the proportion of each hour.
Simple but effective.
}
