% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityoverlap.R
\name{community_overlap}
\alias{community_overlap}
\title{Median pairwise overlap across all species in a community}
\usage{
community_overlap(traits, sp, norm = TRUE, bw = NULL, n = NULL)
}
\description{
Input: vector of traits, vector of species identities
}
