% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_overlap_merged.R
\name{community_overlap_merged}
\alias{community_overlap_merged}
\title{Community Overlap Calculation}
\usage{
community_overlap_merged(
  traits,
  sp,
  data_type = "linear",
  normal = TRUE,
  output = "median",
  weight_type = "hmean",
  randomize_weights = FALSE,
  circular_args = list(),
  density_args = list()
)
}
\arguments{
\item{traits}{A numeric vector of trait measurements with length equal to the number of individuals. Each individual should have a trait measurement.}

\item{sp}{A factor vector with length equal to the number of elements in the traits vector that indicates the taxon of each individual.}

\item{data_type}{Data type can be "linear", "circular", or "circular_discrete". The data_type argument defaults to "linear" and is appropriate when the trait data are one dimensional (i.e., body masses of individuals). The alternative inputs for this argument are for circular data that are periodic or discoid (e.g., measured in radians or degrees). The argument "circular" is for continuous data and the argument "circular_discrete" is for non-continuous, hourly data.}

\item{normal}{If TRUE, then the area under all density functions is normalized to one. TRUE is the default for this argument.
If FALSE, then the area under all density functions is proportional to the number of observations in that group.}

\item{output}{The arguments *output* specifies the whether the median or mean of all pairwise overlap values between distributions of species will be returned. The default value for output is "median".}

\item{weight_type}{Specifies weights to be used to calculate the median or mean output. When the default value for *weight_type* = "hmean", the function calculates the pairwise overlaps of trait distributions of all species in each community as 2/(1/abundance of species a + 1/abundance of species b). The harmonic mean *weight_type* is set as the default as it minimizes the effect of outliers and rare species. If the argument weight_type = "none", no weights are used for the calculation of mean or median. If weight_type = "mean", raw abundances of species are used as weights.}

\item{randomize_weights}{If randomize_weights is TRUE, the weights given to pairwise overlaps within a community are shuffled when generating null models. The default is set to FALSE, so that weights of pairwise overlaps are not shuffled.}

\item{circular_args}{Optional list of additional arguments to pass to
\code{\link[circular]{circular}}. Only used if the data_type argument is set to "circular".}

\item{density_args}{Additional arguments to pass to \code{\link[stats]{density}}, such as
\code{bw}, \code{n}, or \code{adjust}. If none are provided, default values
of bw = "nrd0", adjust = 1, and n = 512 are used.}
}
\value{
The function returns overall species overlap for a community. At default, it returns the
median of pairwise overlaps weighted by harmonic means of abundances for the community.
}
\description{
This function calculates the median or mean of pairwise overlaps between density
estimates of trait distributions of all species within a community, which can
be weighted by species abundances.
}
\details{
The function evaluates weighted mean or median of overlaps of density
estimates of all species (with abundances >1) in a community. The default calculates the median of pairwise overlaps for the whole community using the harmonic means of abundances of the species pairs as
weights, which minimizes the effect of outliers and rare species. If the argument
weight_type = "none", no weights are used for the calculation of the mean or median. If
weight_type = "mean", arithmetic means of abundances are used as weights. To change the
output to mean, rather than the default output of median, specify the argument output = "mean".
}
\examples{
# Load necessary packages to run the example.
library(tidyverse)
library(Ostats)

# Load small mammal data from Read et al. (2018) and subset data for Harvard Forest (HARV), selecting columns for site, taxon, and body weight (g). Create a new column that is the log10 transformation of body weight.
dat <- read_csv('https://ndownloader.figshare.com/files/9167548') \%>\%
   filter(siteID \%in\% 'HARV') \%>\%
   select(siteID, taxonID, weight) \%>\%
   filter(!is.na(weight)) \%>\%
   mutate(log_weight = log10(weight))

# Calculate median of pairwise overlaps for the community, weighted by harmonic means
of abundances
community_overlap_merged(traits = as.matrix(dat$log_weight),
   sp = factor(dat$taxonID))

}
\references{
Read, Q. D. et al. Among-species overlap in rodent body size
  distributions predicts species richness along a temperature gradient.
  Ecography 41, 1718-1727 (2018).
}
\seealso{
\code{\link{pairwise_overlap}} To calculate overlap between two empirical
density estimates.

\code{\link{circular_overlap}} To calculate continous circular overlap between
two empirical density estimates.

\code{\link{circular_overlap_24hour}} To calculate overlap for discrete hourly
data.
}
