% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_overlap_merged.R
\name{community_overlap_merged}
\alias{community_overlap_merged}
\title{Community Overlap Calculation}
\usage{
community_overlap_merged(
  traits,
  sp,
  normal = TRUE,
  output = "median",
  weight_type = "hmean",
  bw = NULL,
  N = NULL,
  itv = TRUE,
  randomize_weights = FALSE,
  display = FALSE
)
}
\arguments{
\item{traits}{a vector of trait measurement.}

\item{sp}{a vector with length equal to length(traits) that indicates the
taxon of each individual.}

\item{normal}{If TRUE, assume data are normally distributed; if FALSE,
additional normalization step is carried out by multiplying each density 
entry by the length of each vector.}

\item{output}{specifies whether median or mean is calculated.}

\item{weight_type}{specifies weights to be used to calculate the median or mean.}

\item{bw}{the smoothing bandwidth to be used. The kernels are scaled such
that this is the standard deviation of the smoothing kernel.}

\item{N}{the number of equally spaced points at which the density is to be
estimated.}

\item{itv}{If TRUE, pairwise overlaps are evaluated;if FALSE, pairwise distances
are calculated, which means each species is reduced to a trait mean and the 
intraspecific trait variation (ITV) is ignored.}

\item{randomize_weights}{If TRUE, randomize weights given to pairwise overlaps
within a community. This can be used to generate null models.}

\item{display}{If TRUE, display all pairwise overlaps.}
}
\value{
At default, the function returns a median of pairwise overlaps weighted by 
harmonic means of abundances for the community.
}
\description{
This function calculates the median or mean of pairwise overlaps between density 
estimates of trait distributions of all species within a community, which can 
be weighted by species abundances. It can also evaluate pairwise distances 
calculated from trait means.
}
\details{
The funtion evaluates weighted pairwise distances or overlaps of density 
estimates of all species in a community taking complete cases with species abundances 
greater than 1 from the dataset. The default calculates the median of pairwise overlaps
for the whole community using the harmonic means of abundances of the species pairs as 
weights, which minimizes the effect of outliners.If the argument weights == NULL, no 
weights are used for the calculation of mean/median. If weights == "mean", means of 
abundances are used as weights. To change the output to mean, specify the argument 
output == "mean".
}
\note{
The function so far only supports overlap statistics for one trait.
}
\examples{
library(tidyverse)
library(Ostats)

# Load data from web archive and Keep only the relevant part of data
dat <- read_csv('https://ndownloader.figshare.com/files/9167548') \%>\%
   filter(siteID \%in\% 'HARV') \%>\%
   select(siteID, taxonID, weight) \%>\%
   filter(!is.na(weight)) \%>\%
   mutate(log_weight = log10(weight))

# Calculate median of pairwise overlaps for the community,weighted by harmonic median
of abundances
community_overlap_merged(traits = as.matrix(dat$log_weight), 
sp = factor(dat$taxonID))
}
\references{
Read, Q. D. et al. Among-species overlap in rodent body size
  distributions predicts species richness along a temperature gradient.
  Ecography 41, 1718-1727 (2018).
}
\seealso{
\code{\link{pairwise_overlap}} to calculate overlap between two empirical 
density estimates.
}
