% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_overlap_merged.R
\name{community_overlap_merged}
\alias{community_overlap_merged}
\title{Community Overlap Calculation}
\usage{
community_overlap_merged(
  traits,
  sp,
  data_type = "linear",
  normal = TRUE,
  output = "median",
  weight_type = "hmean",
  randomize_weights = FALSE,
  circular_args = list(),
  density_args = list()
)
}
\arguments{
\item{traits}{a vector of trait measurement.}

\item{sp}{a vector with length equal to length(traits) that indicates the
taxon of each individual.}

\item{data_type}{data type can be "linear", "circular",or "circular_discrete".
Default to "linear".}

\item{normal}{if TRUE, the area under all density functions is normalized to 1,
if FALSE, the area under all density functions is proportional to the number of
observations in that group.}

\item{output}{specifies whether median or mean is calculated.}

\item{weight_type}{specifies weights to be used to calculate the median or mean.}

\item{randomize_weights}{If TRUE, randomize weights given to pairwise overlaps
within a community. This can be used to generate null models.}

\item{circular_args}{list of additional arguments to be passed to
\code{\link[circular]{circular}}. Only used if \code{data_type} is "circular".}

\item{density_args}{list of additional arguments to be passed to
\code{\link[stats]{density}}.}
}
\value{
The function returns overall species overlap for a community. At default, it returns the
median of pairwise overlaps weighted by harmonic means of abundances for the community.
}
\description{
This function calculates the median or mean of pairwise overlaps between density
estimates of trait distributions of all species within a community, which can
be weighted by species abundances.
}
\details{
The function evaluates weighted mean or median of overlaps of density
estimates of all species in a community taking complete cases with species abundances
greater than 1 from the dataset. The default calculates the median of pairwise overlaps
for the whole community using the harmonic means of abundances of the species pairs as
weights, which minimizes the effect of outliners and rare species.If the argument
weight_type = "none", no weights are used for the calculation of mean/median. If
weight_type = "mean", arithmetic means of abundances are used as weights. To change the
output to mean, specify the argument output = "mean".
}
\examples{
library(Ostats)

# Keep only the relevant part of small mammal data
dat <- small_mammal_data[small_mammal_data$siteID \%in\% c('HARV','JORN'), ]
dat <- dat[!is.na(dat$weight), ]
dat$log_weight <- log10(dat$weight)

# Calculate median of pairwise overlaps for the community,weighted by harmonic means
# of abundances
community_overlap_merged(traits = as.matrix(dat$log_weight),
   sp = factor(dat$taxonID))

}
\references{
Read, Q. D. et al. Among-species overlap in rodent body size
  distributions predicts species richness along a temperature gradient.
  Ecography 41, 1718-1727 (2018).
}
\seealso{
\code{\link{pairwise_overlap}} to calculate overlap between two empirical
density estimates.
}
