% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityoverlap.R
\name{community_overlap_noitv}
\alias{community_overlap_noitv}
\title{Abundance-weighted median pairwise distance between species trait means in a community}
\usage{
community_overlap_noitv(traits, sp, norm = TRUE, bw = NULL, n = NULL)
}
\description{
Input: vector of traits, vector of species identities
}
\details{
This reduces each species to a trait mean so that you can compare the inference from overlap to a
situation where intraspecific trait variation (ITV) is ignored.
}
