% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityoverlap.R
\name{community_overlap_wmedian}
\alias{community_overlap_wmedian}
\title{Abundance-weighted Median of Pairwise Overlaps across All species in a Community}
\usage{
community_overlap_wmedian(traits, sp, norm = TRUE, bw = NULL, n = NULL)
}
\arguments{
\item{traits}{a vector of trait measurement.}

\item{sp}{a vector with length equal to length(traits) that indicates the
taxon of each individual.}

\item{norm}{If TRUE, assume data are normally distributed; if FALSE,
additional normalization step is carried out by multiplying each density 
entry by the length of each vector.}

\item{bw}{the smoothing bandwidth to be used. The kernels are scaled such
that this is the standard deviation of the smoothing kernel.}

\item{n}{the number of equally spaced points at which the density is to be
estimated.}
}
\value{
The function returns a median of pairwise overlaps of all species in 
the community, weighted by means of abundances of species pairs.
}
\description{
This function calculates the median of pairwise overlaps between density 
estimates of trait distributions of all species within a community, weighted by 
the mean of abundances of the species pairs.
}
\details{
The funtion evaluates pairwise overlaps of density estimates of all 
species in a community taking complete cases with species abundances greater 
than 1 from the original dataset, using the mean of abundances of species 
pairs as weights to calculate the weighted median of overlaps for the community.
}
\seealso{
\code{\link{pairwise_overlap}} to calculate overlap between two empirical 
density estimates.

\code{\link{community_overlap_harmonicwmedian}} to see a function that
uses harmonic means of abundances of species pairs as weights to calculate median 
of overlaps.
}
