% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostat2longform.R
\name{Ostat2longform}
\alias{Ostat2longform}
\title{Converts O-stats output to long format}
\usage{
Ostat2longform(o)
}
\arguments{
\item{o}{The output of a call to \code{Ostats}.}
}
\value{
A dataframe with the number of rows equal to the number of communities (i.e., plots) that were evaluated by the Ostats function and the number of columns equals to  nrows = the names of communities and the number of columns equal to 14 (i.e., one for site name, one for trait name, and 12 for the outputs of the Ostats function).
}
\description{
This function converts the output of Ostats to long format for subsequent data analysis. For an example, see Read et al. (2018) wherein outputs of the Ostats function were converted to long format to perform a path analysis.
}
\examples{
# Load data from Read et al. (2018) from Figshare web archive
dat <- read_csv('https://ndownloader.figshare.com/files/9167548')

# Keep only sites "HARV" and "JORN" using the filter function, and select the relevant columns required  # by the function.
# Use the mutate function to add a new column named "log_weight" to log-transform the measurements.

dat <- dat \%>\%
   filter(siteID \%in\% c('HARV','JORN')) \%>\%
   select(siteID, taxonID, weight) \%>\%
   filter(!is.na(weight)) \%>\%
   mutate(log_weight = log10(weight))

# Group the data by siteID and taxonID and look at the summary
dat \%>\%
   group_by(siteID, taxonID) \%>\%
   slice(1)

# Run O-stats on the small mammal data with all arguments set to default
 Ostats_example <- Ostats(traits = as.matrix(dat[,'log_weight']),
                    sp = factor(dat$taxonID),
                    plots = factor(dat$siteID),
                    data_type = "linear")

# Convert the output generated by the Ostats function to long format
Ostat2longform(Ostats_example)
}
\references{
Read, Q.D., J.M. Grady, P.L. Zarnetske, S. Record, B. Baiser, J. Belmaker, M.-N. Tuanmu, A. Strecker, L. Beaudrot, and K.M. Thibault. 2018. Among-species overlap in rodent body size distributions predicts species richness along a temperature gradient. *Ecography* 41(10):1718-1727. https://doi.org/10.1111/ecog.03641
}
\seealso{
\code{\link{Ostats}} To evaluate against a local null model.

