% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats.R
\name{Ostats}
\alias{Ostats}
\title{Calculate O-statistics (community-level pairwise niche overlap statistics)}
\usage{
Ostats(
  traits,
  plots,
  sp,
  nperm = 99,
  nullqs = c(0.025, 0.975),
  shuffle_weights = FALSE,
  swap_means = FALSE
)
}
\arguments{
\item{traits}{a matrix dataset with nrows = n individuals, ncols = n traits.}

\item{plots}{a factor with length equal to nrow(traits) that indicates the
community each individual belongs to.}

\item{sp}{a factor with length equal to nrow(traits) that indicates the taxon
of each individual.}

\item{nperm}{the number of permutations to generate a null model.}

\item{nullqs}{numeric vector of probabilities with values in [0,1] to set
effect size quantiles.}

\item{shuffle_weights}{If TRUE, shuffle weights given to pairwise overlaps
within a community when generating null models.}

\item{swap_means}{If TRUE, swap means of body sizes within a community when
generating null models.}
}
\value{
The funtion returns a list containing 4 objects: overlaps_norm, which
  is the median of weighted pairwise overlaps of trait distributions of all
  species in each community, assuming normal distribution. overlaps_unnorm,
  the medians calculated assuming kernel densities are not normalized.
  overlaps_norm_ses, effect size statistics against a null model assuming
  normal distribution, and overlaps_unnorm_ses, effect size statistics with
  the additional normalization step.
}
\description{
This is the primary function in the Ostats package. It evaluates the
O-statistics against a local null model.
}
\details{
This function evaluates the overlap statistics against a local null
  model.The function community_overlap_harmonicwmedian calculates the median
  of pairwise overlaps of trait distributions of all species in each
  community, weighted by harmonic mean of species abundaces of the species
  pairs in each community. Two results are produced, one assuming the kernel
  densities are normally distributed, and the other assuming the kernel
  densities are not normalized, thus each density entry is multiplied by the
  length of each vector before computing the overlap coefficients.

  When both shuffle_weights and swap_means are FALSE, null models are
  generated by randomly assigning a taxon that is present in the community to
  each individual. The medians of pairwise overlaps are then calculated to be
  compared against.

  Effect size statistics are calculated between the median for each community
  and the null model.
}
\note{
The function so far only supports overlap statistics for one trait.
}
\examples{
# overlap statistics for body weights of species in NEON sites
library(tidyverse)
library(Ostats)
# Load data from web archive
dat <- read_csv('https://ndownloader.figshare.com/files/9167548')

# Keep only the relevant part of data
dat <- dat \%>\%
  filter(siteID \%in\% c('HARV','JORN')) \%>\%
  select(siteID, taxonID, weight) \%>\%
  filter(!is.na(weight)) \%>\%
  mutate(log_weight = log10(weight))

dat \%>\%
  group_by(siteID, taxonID) \%>\%
  slice(1)

#Run O-stats on the data with nperm = 2 (do not bother with null models)
Ostats_example <- Ostats(traits = as.matrix(dat$log_weight),
                   sp = factor(dat$taxonID),
                   plots = factor(dat$siteID),
                   nperm = 2)
}
\references{
Read, Q. D. et al. Among-species overlap in rodent body size
  distributions predicts species richness along a temperature gradient.
  Ecography 41, 1718-1727 (2018).
}
\seealso{
\code{\link{Ostats_regional}} to evaluate against a regional null
  model.

\code{\link{Ostats_circular}} for circular distributions.
}
