% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats.R
\name{Ostats}
\alias{Ostats}
\title{Calculate O-statistics (community-level pairwise niche overlap statistics)}
\usage{
Ostats(
  traits,
  plots,
  sp,
  data_type = "linear",
  output = "median",
  weight_type = "hmean",
  run_null_model = TRUE,
  nperm = 99,
  nullqs = c(0.025, 0.975),
  shuffle_weights = FALSE,
  swap_means = FALSE,
  random_seed = NULL,
  circular_args = list(),
  density_args = list()
)
}
\arguments{
\item{traits}{a vector or matrix of trait measurements. The number of elements
in the vector or number of rows in the matrix is the number of individuals,
and the number of columns of the matrix is the number of traits.}

\item{plots}{a factor with length equal to nrow(traits) that indicates the
community each individual belongs to.}

\item{sp}{a factor with length equal to nrow(traits) that indicates the taxon
of each individual.}

\item{data_type}{data type can be "linear", "circular",or "circular_discrete".
Defaults to "linear".}

\item{output}{specifies whether median or mean is calculated.}

\item{weight_type}{specifies weights to be used to calculate the median or mean.}

\item{run_null_model}{whether to run a null model (if \code{TRUE}) and evaluate the
O-statistics against it, or simply return the raw O-statistics (if \code{FALSE}).
Defaults to \code{TRUE}.}

\item{nperm}{the number of null model permutations to generate. Defaults to 99.}

\item{nullqs}{numeric vector of probabilities with values in [0,1] to set
effect size quantiles. Defaults to \code{c(0.025, 0.975)}.}

\item{shuffle_weights}{If \code{TRUE}, shuffle weights given to pairwise overlaps
within a community when generating null models.}

\item{swap_means}{If \code{TRUE}, swap means of body sizes within a community when
generating null models.}

\item{random_seed}{User may supply a random seed to enable reproducibility
of null model output. A warning is issued, and a random seed is generated
based on the local time, if the user does not supply a seed.}

\item{circular_args}{optional list of additional arguments to pass to
\code{\link[circular]{circular}}. Only used if the data type is "circular".
Note that continuous circular data must be provided in radian units.}

\item{density_args}{additional arguments to pass to \code{\link[stats]{density}}, such as
\code{bw}, \code{n}, or \code{adjust}. If none are provided, default values
are used.}
}
\value{
The function returns a list containing 4 objects:
\item{overlaps_norm}{a matrix showing the median of weighted pairwise overlaps (at default) of trait
  distributions of all species in each community (at default output
  and weight_type), with the area under all density functions normalized to 1.}
\item{overlaps_unnorm}{a matrix showing O-stats calculated with the area under all density
  functions proportional to the number of observations in that group.}
\item{overlaps_norm_ses}{5 matrices of effect size statistics against a null model
  with the area under all density functions normalized to 1.}
\item{overlaps_unnorm_ses}{5 matrices of effect size statistics against a null model
  with the area under all density functions proportional to the number
  of observations in that group.}
}
\description{
This is the primary function in the Ostats package. It evaluates the
O-statistics against a local null model.
}
\details{
This function evaluates the overlap statistics against a local null
  model. By default, it calculates the median of pairwise
  overlaps, weighted by harmonic mean of species abundaces of the species
  pairs in each community. Two results are produced, one normalizing the area
  under all density functions to 1, the other making the area under all density
  functions proportional to the number of observations in that group.

  When both shuffle_weights and swap_means are FALSE, null models are
  generated by randomly assigning a taxon that is present in the community to
  each individual. O-stats are then calculated to be compared against.

  Effect size statistics are calculated between the observed O-stats
  and the null model.
}
\examples{
\dontrun{
# overlap statistics for body weights of small mammals in NEON sites
library(Ostats)

# Keep only the relevant part of data
dat <- small_mammal_data[small_mammal_data$siteID \%in\% c('HARV','JORN'), ]
dat <- dat[!is.na(dat$weight), ]
dat$log_weight <- log10(dat$weight)


#Run O-stats on the data with only a few null model iterations
Ostats_example <- Ostats(traits = as.matrix(dat[,'log_weight']),
                   sp = factor(dat$taxonID), data_type = "linear",
                   plots = factor(dat$siteID),
                   nperm = 10)
}
}
\references{
Read, Q. D. et al. Among-species overlap in rodent body size
  distributions predicts species richness along a temperature gradient.
  Ecography 41, 1718-1727 (2018).
}
\seealso{
\code{\link{Ostats_multivariate}} for multidimensional overlap.

\code{\link{Ostats_plot}} for plotting community overlap for
each community.
#'
}
\author{
Quentin D. Read, John M. Grady, Arya Y. Yue, Isadora Fluck E., Ben Baiser,
Angela Strecker, Phoebe L. Zarnetske, and Sydne Record
}
