% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats.R
\name{Ostats}
\alias{Ostats}
\title{Calculate O-statistics (community-level pairwise niche overlap statistics)}
\usage{
Ostats(
  traits,
  plots,
  sp,
  data_type = "linear",
  output = "median",
  weight_type = "hmean",
  nperm = 99,
  nullqs = c(0.025, 0.975),
  shuffle_weights = FALSE,
  swap_means = FALSE,
  circular_args = list(),
  ...
)
}
\arguments{
\item{traits}{a vector of trait measurement with nrow = n individuals.}

\item{plots}{a factor with length equal to nrow(traits) that indicates the
community each individual belongs to.}

\item{sp}{a factor with length equal to nrow(traits) that indicates the taxon
of each individual.}

\item{data_type}{data type can be "linear", "circular",or "circular_discrete".
Defaults to "linear".}

\item{output}{specifies whether median or mean is calculated.}

\item{weight_type}{specifies weights to be used to calculate the median or mean.}

\item{nperm}{the number of permutations to generate a null model.}

\item{nullqs}{numeric vector of probabilities with values in [0,1] to set
effect size quantiles.}

\item{shuffle_weights}{If TRUE, shuffle weights given to pairwise overlaps
within a community when generating null models.}

\item{swap_means}{If TRUE, swap means of body sizes within a community when
generating null models.}

\item{circular_args}{optional list of additional arguments to pass to
\code{\link[circular]{circular}}. Only used if the data type is "circular".}

\item{...}{additional arguments to pass to \code{\link[stats]{density}}, such as
\code{bw}, \code{n}, or \code{adjust}. If none are provided, default values
are used.}
}
\value{
The funtion returns a list containing 4 objects:
\item{overlaps_norm}{the median of weighted pairwise overlaps of trait
  distributions of all species in each community, with the area
  under all density functions normalized to 1.}
\item{overlaps_unnorm}{the medians calculated with the area under all density
  functions proportional to the number of observations in that group.}
\item{overlaps_norm_ses}{effect size statistics against a null model
  with the area under all density functions normalized to 1.}
\item{overlaps_unnorm_ses}{effect size statistics with the area under all density
  functions proportional to the number of observations in that group.}
}
\description{
This is the primary function in the Ostats package. It evaluates the
O-statistics against a local null model.
}
\details{
This function evaluates the overlap statistics against a local null
  model.The function community_overlap_merged calculates the community overlap
  statistics in each community. At default, it calculates the median of pairwise
  overlaps, weighted by harmonic mean of species abundaces of the species
  pairs in each community. Two results are produced, one normalizing the area
  under all density functions to 1, the other making the area under all density
  functions proportional to the number of observations in that group.

  When both shuffle_weights and swap_means are FALSE, null models are
  generated by randomly assigning a taxon that is present in the community to
  each individual. O-stats are then calculated to be compared against.

  Effect size statistics are calculated between the observed O-stats
  and the null model.
}
\note{
The function so far only supports overlap statistics for one trait.
}
\examples{
# overlap statistics for body weights of species in NEON sites
library(tidyverse)
library(Ostats)
# Load data from web archive
dat <- read_csv('https://ndownloader.figshare.com/files/9167548')

# Keep only the relevant part of data
dat <- dat \%>\%
  filter(siteID \%in\% c('HARV','JORN')) \%>\%
  select(siteID, taxonID, weight) \%>\%
  filter(!is.na(weight)) \%>\%
  mutate(log_weight = log10(weight))

dat \%>\%
  group_by(siteID, taxonID) \%>\%
  slice(1)

#Run O-stats on the data with nperm = 2 (do not bother with null models)
Ostats_example <- Ostats(traits = as.matrix(dat$log_weight),
                   sp = factor(dat$taxonID), data_type = "linear",
                   plots = factor(dat$siteID),
                   nperm = 2)
}
\references{
Read, Q. D. et al. Among-species overlap in rodent body size
  distributions predicts species richness along a temperature gradient.
  Ecography 41, 1718-1727 (2018).
}
\seealso{
\code{\link{Ostats_regional}} to evaluate against a regional null
  model.

\code{\link{Ostats_circular}} for circular distributions.

\code{\link{community_overlap_merged}} for median of
pairwise overlaps of species trait distributions within a community.

\code{\link{get_ses}} for standardized effect sizes from null
model values.
}
\author{
Quentin D. Read, John M. Grady, Arya Y. Yue, Isadora Fluck E., Ben Baiser,
Angela Strecker, Phoebe L. Zarnetske, and Sydne Record
}
