% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats.R
\name{Ostats}
\alias{Ostats}
\title{Calculate O-statistics (community-level pairwise niche overlap statistics)}
\usage{
Ostats(traits, plots, sp, nperm = 99, nullqs = c(0.025, 0.975),
  shuffle_weights = FALSE, swap_means = FALSE)
}
\description{
This is the primary function in the Ostats package. It evaluates the
O-statistics against a local null model.
}
\seealso{
\code{\link{Ostats_regional}} to evaluate against a regional null model.
}
