% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats_plot.R
\name{Ostats_plot}
\alias{Ostats_plot}
\title{Plotting Community Overlap}
\usage{
Ostats_plot(indiv_dat, siteID, taxonID, trait, overlap_dat, sites2use = NULL)
}
\arguments{
\item{indiv_dat}{Individual data: a data frame containing individual measurments for a certain
traits,its species identity, and the community identity it belongs to.}

\item{siteID}{Site identity: a column in indiv_dat data frame that indicates the names of each community.}

\item{taxonID}{Taxon identity: a column in indiv_dat data frame that indicates species or taxa names.}

\item{trait}{The trait you want to overlap among species: a column in indiv_dat data frame containing trait measurements for each individual.}

\item{overlap_dat}{an object containing the results from the Ostats function - see \code{\link{Ostats}}}

\item{sites2use}{a vector that select the sites you want to plot. If NULL, the function will plot all the sites.}
}
\value{
Density plots of species trait distribution plotted on the same graph
 for each community to show how they overlap each other. The overlap value obtained as output from the function \code{\link{Ostats}}, is labelled on each community graph.
}
\description{
This function plots the overlap of traits among
 species for each community.
}
\examples{
library(tidyverse)
library(ggplot2)
library(dplyr)

overlap_dat <- Ostats_example #your results from the Ostat function - see \code{\link{Ostats}}
indiv_dat <- read_csv('https://ndownloader.figshare.com/files/9167548', col_names = T)
siteID <- indiv_dat$siteID
taxonID <- indiv_dat$taxonID
trait <- indiv_dat$logweight
#to plot only selected sites:
sites2use<- c('BART','KONZ','JORN') #used in the example

#to plot all sites:
sites2use<- NULL
Ostats_plot(indiv_dat = indiv_dat, siteID = siteID, taxonID = taxonID, trait = trait, overlap_dat = overlap_dat, sites2use = sites2use)
}
\seealso{
\code{\link{Ostats}} to Calculate O-statistics (community-level
 pairwise niche overlap statistics)
}
