% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats_plot.R
\name{Ostats_plot}
\alias{Ostats_plot}
\title{Plotting Community Overlap}
\usage{
Ostats_plot(indiv_dat, siteID, taxonID, trait, overlap_dat, trait_name)
}
\arguments{
\item{indiv_dat}{a data frame containing individual measurments for a certain
trait,its species identity, and the community it belongs to.}

\item{siteID}{a column in indiv_dat that indicates the community.}

\item{taxonID}{a column in indiv_dat that indicates species identity.}

\item{trait}{a column of trait measurements in indiv_dat.}

\item{overlap_dat}{a data frame containing .. NEED TO FIGURE OUT HOW TO USE
OVERLAP OUTPUT}

\item{trait_name}{a character that specifies the name of the trait, to be
labelled on the x axis.}
}
\value{
Density plots of species trait distribution are plotted on the same graph
 for each community to show how they overlap with each other. With the output from
 the function \code{\link{Ostats}}, the overlap value is labelled on each graph.
}
\description{
This function plots the overlap of trait measurements of among
 species for each community.
}
\examples{

#FUNCTION NOT COMPLETE!!!
#example use same dataset as ostats (need to change later)

library(tidyverse)
overlap_dat <- read.csv('overlap_dat.csv')
indiv_dat <- read.csv('indiv_dat.csv')
indiv_dat <- indiv_dat \%>\%filter(siteID \%in\% c('HARV','JORN'))
overlap_dat <- overlap_dat \%>\%filter(siteID \%in\% c('HARV','JORN'))

indiv_dat <- mutate(indiv_dat, log_weight = log10(weight))

#change column name to test whether the function works
names(indiv_dat) <- c("siteID", "species","weight", "log_weight")

Ostats_plot(indiv_dat = indiv_dat,siteID=indiv_dat$siteID, taxonID = indiv_dat$species, trait = indiv_dat$log_weight,overlap_dat = overlap_dat, trait_name = "log_weight")
}
\seealso{
\code{\link{Ostats}} to Calculate O-statistics (community-level
 pairwise niche overlap statistics)
}
