% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats_plot.R
\name{Ostats_plot}
\alias{Ostats_plot}
\title{Plotting Community Overlap}
\usage{
Ostats_plot(indiv_dat, siteID, taxonID, trait, overlap_dat, trait_name)
}
\arguments{
\item{indiv_dat}{Individual data: a data frame containing individual measurments for a certain
traits,its species identity, and the community identity it belongs to.}

\item{siteID}{Site identity: a column in indiv_dat data frame that indicates the names of each community.}

\item{taxonID}{Taxon identity: a column in indiv_dat data frame that indicates species or taxa names.}

\item{trait}{The trait you want to overlap among species: a column in indiv_dat data frame containing trait measurements for each individual.}

\item{overlap_dat}{a data frame where the first column contains the names of each community and the second column contains the values of total overlap among species in each of these communities obtained from the \code{\link{Ostats}} function  NEED TO FIGURE OUT HOW TO USE OVERLAP OUTPUT}

\item{trait_name}{a character that specifies the name of the trait, to be
labelled on the x axis.}
}
\value{
Density plots of species trait distribution plotted on the same graph
 for each community to show how they overlap each other. The overlap value obtained as output from the function \code{\link{Ostats}}, is labelled on each community graph.
}
\description{
This function plots the overlap of traits among
 species for each community.
}
\examples{

#FUNCTION NOT COMPLETE!!!
#example use same dataset as ostats (need to change later)

library(tidyverse)
overlap_dat <- read.csv('overlap_dat.csv')
indiv_dat <- read.csv('indiv_dat.csv')
indiv_dat <- indiv_dat \%>\%filter(siteID \%in\% c('HARV','JORN'))
overlap_dat <- overlap_dat \%>\%filter(siteID \%in\% c('HARV','JORN'))

indiv_dat <- mutate(indiv_dat, log_weight = log10(weight))

#change column name to test whether the function works
names(indiv_dat) <- c("siteID", "species","weight", "log_weight")

Ostats_plot(indiv_dat = indiv_dat,siteID=indiv_dat$siteID, taxonID = indiv_dat$species, trait = indiv_dat$log_weight,overlap_dat = overlap_dat, trait_name = "log_weight")
}
\seealso{
\code{\link{Ostats}} to Calculate O-statistics (community-level
 pairwise niche overlap statistics)
}
