% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats_plot.R
\name{Ostats_plot}
\alias{Ostats_plot}
\title{Plotting Community Overlap}
\usage{
Ostats_plot(
  indiv_dat,
  siteID,
  taxonID,
  trait,
  overlap_dat,
  sites2use = NULL,
  n_col = 3,
  colorvalues = NULL,
  alpha_o = 0.5,
  adjust_o = 2,
  limits_o = c(0.5 * min(trait, na.rm = TRUE), 1.5 * max(trait, na.rm = TRUE)),
  name_x = "Trait value",
  name_y = "Probability Density",
  media = FALSE
)
}
\arguments{
\item{indiv_dat}{Individual data: a data frame containing individual measurments for a certain
traits,its species identity, and the community identity it belongs to.}

\item{siteID}{Site identity: a column in indiv_dat data frame that indicates the names of each community.}

\item{taxonID}{Taxon identity: a column in indiv_dat data frame that indicates species or taxa names.}

\item{trait}{The trait you want to overlap among species: a column in indiv_dat data frame containing trait measurements for each individual.}

\item{overlap_dat}{an object containing the results from the Ostats function - see \code{\link{Ostats}}}

\item{sites2use}{a vector that select the sites you want to plot. If NULL, the function will plot all the sites.}

\item{n_col}{Number of columns for layout of individual panels. Default is 3.}

\item{colorvalues}{Vector of color values for the density polygons. Defaults to a viridis palette if none provided.}

\item{alpha_o}{defines the colors trasparency level for the density polygons. Default is 0.5}

\item{adjust_o}{multiplicate the bandwidth adjustment of the density polygons. The less, the tiny your density polygons will be. Default is 2.}

\item{limits_o}{the limits (min and max values) of the x axis. Default is \code{c(0.5*min(trait,na.rm=TRUE), 1.5*max(trait,na.rm=TRUE))}}

\item{name_x}{a character indicating the name of your x axis (i.e. the name of your trait). Default is 'Trait value'}

\item{name_y}{a character indicating the name of your y axis. Default is 'Probability Density'}

\item{media}{if TRUE it plots the trait media for each species. Default is FALSE.}
}
\value{
Density plots of species trait distribution plotted on the same graph
 for each community to show how they overlap each other. The overlap value obtained as output from the function \code{\link{Ostats}}, is labelled on each community graph.
}
\description{
This function plots the overlap of traits among
 species for each community.
}
\examples{
library(tidyverse)
library(Ostats)

indiv_dat <- read_csv('https://ndownloader.figshare.com/files/9167548', col_names = T)

Ostats_plot(indiv_dat = indiv_dat, siteID = indiv_dat$siteID, taxonID = indiv_dat$taxonID, trait = indiv_dat$logweight, overlap_dat = Ostats_example, sites2use = c('HARV','JORN'))
}
\seealso{
\code{\link{Ostats}} to Calculate O-statistics (community-level
 pairwise niche overlap statistics)
}
