% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats.R
\name{Ostats_regional}
\alias{Ostats_regional}
\title{Calculate O-statistics evaluated against regional null model}
\usage{
Ostats_regional(
  traits,
  plots,
  sp,
  reg_pool_traits,
  reg_pool_sp,
  nperm = 99,
  nullqs = c(0.025, 0.975),
  density_args = list()
)
}
\arguments{
\item{traits}{A numeric vector of trait measurements with length equal to the number of individuals in the focal community to be compared against the regional null model. Each individual should have a trait measurement.}

\item{plots}{A factor vector with length equal to the number of elements in the traits vector that indicates the community in which each individual belongs.}

\item{sp}{A factor vector with length equal to the number of elements in the traits vector that indicates the taxon of each individual.}

\item{reg_pool_traits}{A list of one-dimensional matrices containing individual trait measurements from different communities in space and/or time that will make up the regional pool of traits which will be compared to the traits of the focal community. The number of items in the list corresponds with the number of unique sites in space that the focal community will be compared to.}

\item{reg_pool_sp}{A list of one-dimensional matrices containing taxonomic names of individuals from the different communities in space and/or time that will make up the regional pool of species whose traits which will be compared to the traits of the focal community. Note that the number of rows in each matrix contained in this argument needs to equal the number of rows in each matrix contained in the reg_pool_traits argument (i.e., each list item in reg_pool_traits needs to have the same dimension as each list item in reg_pool_sp). The number of items in the list corresponds with the number of unique sites in space that the focal community will be compared to.}

\item{nperm}{The number of permutations,or randomly generated data null community subsets, to generate a null model. The default is set to 99 permutations.}

\item{nullqs}{A numeric vector of length two containing the lower and upper effect size quantiles for the null models. The default lower and upper effect size quantiles, respectively, are 0.025 and 0.975. The values in this vector are probabilities, so they are bounded between zero and one.}

\item{density_args}{Additional arguments to pass to \code{\link[stats]{density}}, such as
\code{bw}, \code{n}, or \code{adjust}. If none are provided, default values
of bw = "nrd0", adjust = 1, and n = 512 are used.}
}
\value{
The function returns a list containing 3 items:
  \item{overlaps_reg}{A list containing one item, which is the raw (i.e., not standardized) overlap between the focal community's trait
  distribution and the trait distribution of the regional species pool.}
  \item{overlaps_reg_allpool_ses}{A list containing five items resulting from the total-pool null model. The total-pool null model is         generated by sampling n trait measurements (n = number of trait measurements for the community) from the regional pool of traits measurements and calculates the pairwise overlap between the random sample and the regional pool. The first three items in this list contain the standardized effect size statistics for the total-pool null model and the last two items contain the upper and lower quantiles of the raw overlap value of the total-pool null model.}
  \item{overlaps_reg_bysp_ses}{A list containing five items resulting from the by-species null model. The by-species null model is calculated by sampling trait measurements of the same species from the regional pool for each species present in the community to calculate pairwise overlap between the sample and the whole regional pool. The first three items in this list contain the standardized effect size statistics for the by-species null model and the last two items contain the upper and lower quantiles of the raw overlap value for the by-species null model.}
}
\description{
This function evaluates the O-statistics (community-level pairwise niche
overlap statistics) against a regional species pool null model. In contrast to
\code{Ostats}, it requires trait and species information about the regional
species pool.
}
\details{
This function evaluates the overlap statistics against a regional
  null model. It takes all individual trait measurements in the community to
  compare with the regional pool (ignoring species identity) to calculate
  pairwise overlap for each community, which indicates the regional pool
  space a community takes up.

  Total pool null model is generated by sampling n trait measurements (n =
  number of trait measurements for the community) from the regional pool and
  calculate the pairwise_overlap between the random sample and the regional
  pool. By-species null model is calculated by sampling trait measurements of
  the same species from the regional pool for each species present in the
  community to calculate pairwise overlap between the sample and the whole
  regional pool.

  In this function, the area under all density functions are all normalized to one.
}
\examples{
library(tidyverse)

# Use the 2015 data from the two sites HARV and BART as the focal community
sites2015 <- reg_pool \%>\%
filter(siteID=='HARV'|siteID == 'BART') \%>\%
filter(substring(as.character(endDate),1,4)=="2015")

# Making the overall data (HARV and BART, sites from the same domain) the regional pool across space and time
reg_pool_traits <- list (as.matrix(reg_pool$log_weight), as.matrix(reg_pool$log_weight))
reg_pool_sp <- list(as.matrix(reg_pool$taxonID), as.matrix(reg_pool$taxonID))
names(reg_pool_traits) <- c("HARV", "BART")
names(reg_pool_sp) <- c("HARV", "BART")
Ostats_reg_example <- Ostats_regional(traits = as.matrix(sites2015$log_weight),
                                     plots = factor(sites2015$siteID),
                                     sp = factor(sites2015$taxonID),
                                     reg_pool_traits = reg_pool_traits,
                                     reg_pool_sp = reg_pool_sp,
                                     nperm = 2)

}
\seealso{
\code{\link{Ostats}} To evaluate against a local null model.

\code{\link{pairwise_overlap}} To calculate overlap between two
  empirical density estimates.

\code{\link{get_ses}} To obtain standardized effect sizes from null model
  values.
}
\author{
Quentin Read, John Grady, Arya Y. Yue, Isadora Fluck E., Ben Baiser,
  Angela Strecker, Phoebe Zarnetske, and Sydne Record
}
