% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats.R
\name{Ostats_regional}
\alias{Ostats_regional}
\title{Calculate O-statistics (community-level pairwise niche overlap statistics) evaluated against regional null model}
\usage{
Ostats_regional(
  traits,
  plots,
  sp,
  reg_pool_traits,
  reg_pool_sp,
  nperm = 99,
  nullqs = c(0.025, 0.975)
)
}
\description{
This function evaluates the O-statistics against a regional null model.
In contrast to \code{Ostats}, it requires trait and species information
about the regional species pool.
}
\seealso{
\code{\link{Ostats}} to evaluate against a local null model.
}
