% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ostats.R
\name{Ostats_regional}
\alias{Ostats_regional}
\title{Calculate O-statistics evaluated against regional null model}
\usage{
Ostats_regional(
  traits,
  plots,
  sp,
  reg_pool_traits,
  reg_pool_sp,
  nperm = 99,
  nullqs = c(0.025, 0.975),
  random_seed = NULL,
  density_args = list()
)
}
\arguments{
\item{traits}{a vector of trait measurement with nrow = n individuals.}

\item{plots}{a factor with length equal to nrow(traits) that indicates the
community each individual belongs to.}

\item{sp}{a factor with length equal to nrow(traits) that indicates the taxon
of each individual.}

\item{reg_pool_traits}{trait measurements in the species regional pool to
compare the dataset with.}

\item{reg_pool_sp}{species identification for each measurement in the
regional species pool.}

\item{nperm}{the number of permutations to generate a null model.}

\item{nullqs}{numeric vector of probabilities with values in [0,1] to set
effect size quantiles.}

\item{random_seed}{User may supply a random seed to enable reproducibility
of null model output. A warning is issued, and a random seed is generated
based on the local time, if the user does not supply a seed.}

\item{density_args}{additional arguments to pass to \code{\link[stats]{density}}, such
as \code{bw}, \code{n}, or \code{adjust}. If none are provided, default
values are used.}
}
\value{
The funtion returns a list containing 4 objects:
  \item{overlaps_reg}{pairwise overlap output between the trait data and the
  regional species pool.}
  \item{overlaps_reg_allpool_ses}{effect size
  statistics between observed O-stats and the total-pool null model}
  \item{overlaps_reg_bysp_ses}{effect size statistics between observed
  O-stats and the by-species null model}
}
\description{
This function evaluates the O-statistics (community-level pairwise niche
overlap statistics)against a regional null model. In contrast to
\code{Ostats}, it requires trait and species information about the regional
species pool.
}
\details{
This function evaluates the overlap statistics against a regional
  null model. It takes all individual trait measurements in the community to
  compare with the regional pool (ignoring species identity) to calculate
  pairwise overlap for each community, which indicates the regional pool
  space a community takes up.

  Total pool null model is generated by sampling n trait measurements (n =
  number of trait measurements for the community) from the regional pool and
  calculate the pairwise_overlap between the random sample and the regional
  pool. By-species null model is calculated by sampling trait measurements of
  the same species from the regional pool for each species present in the
  community to calculate pairwise overlap between the sample and the whole
  regional pool.

  In this function, the area under all density functions are normalized to 1.
}
\examples{
# use the 2015 data from the two communities HARV and BART
sites2015 <- reg_pool[reg_pool$siteID \%in\% c('HARV', 'BART'), ]
sites2015 <- sites2015[substr(as.character(sites2015$endDate),1,4)=="2015", ]

# making the overall data (HARV and BART, sites from the same domain)
# the regional pool across space and time
reg_pool_traits <- list (as.matrix(reg_pool$log_weight), as.matrix(reg_pool$log_weight))
reg_pool_sp <- list(as.matrix(reg_pool$taxonID), as.matrix(reg_pool$taxonID))
names(reg_pool_traits) <- c("HARV", "BART")
names(reg_pool_sp) <- c("HARV", "BART")
Ostats_reg_example <- Ostats_regional(traits = as.matrix(sites2015$log_weight),
                                     plots = factor(sites2015$siteID),
                                     sp = factor(sites2015$taxonID),
                                     reg_pool_traits = reg_pool_traits,
                                     reg_pool_sp = reg_pool_sp,
                                     nperm = 2)

}
\seealso{
\code{\link{Ostats}} to evaluate against a local null model.

\code{\link{pairwise_overlap}} to calculate overlap between two
  empirical density estimates.
}
\author{
Quentin Read, John Grady, Arya Y. Yue, Isadora Fluck E., Ben Baiser,
  Angela Strecker, Phoebe Zarnetske, and Sydne Record
}
