% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityoverlap.R
\name{pairwise_overlap}
\alias{pairwise_overlap}
\title{Overlap of two empirical density estimates}
\usage{
pairwise_overlap(a, b, norm = TRUE, bw = NULL, n = NULL)
}
\description{
See \url{http://stats.stackexchange.com/questions/97596/how-to-calculate-overlap-between-empirical-probability-densities},
this function is derived from code posted in the answer by user mmk.
}
