% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_overlap.R
\name{pairwise_overlap}
\alias{pairwise_overlap}
\title{Overlap between Two Empirical Density Estimates}
\usage{
pairwise_overlap(a, b, normal = TRUE, density_args = list())
}
\arguments{
\item{a}{a vector dataset with nrows= n individuals, with a column containing
one measurement of a certain trait.}

\item{b}{another matrix dataset with the same trait measurements to be compared
against a.}

\item{normal}{if TRUE, the area under all density functions is normalized to 1,
if FALSE, the area under all density functions is proportional to the number of
observations in that group.}

\item{density_args}{list of additional arguments to be passed to
\code{\link[stats]{density}} in the univariate case, or
\code{\link[hypervolume]{hypervolume}} in the multivariate case.}
}
\value{
The funtion returns a vector of three values:
\item{overlap_average}{the average overlap of a and b, calculated by the overlap
area *2 divided by the sum of areas under the two functions.}
\item{overlap_a}{the proportion of a that overlaps with b, calculated by the overlap
area divided by area under the function generated from a.}
\item{overlap_b}{the proportion of b that overlaps with a, calculated by the overlap
area diveided by area under the function generated from b.}
}
\description{
This function generates kernel density estimates for two datasets on a common
grid to calculate the area of overlap between the two estimates.
See \url{http://stats.stackexchange.com/questions/97596/how-to-calculate-overlap-between-empirical-probability-densities},
this function is derived from code posted in the answer by user mmk.
}
\details{
This function generates kernel density estimates using the
\code{\link[stats]{density}} function for two datasets on a common grid.
Default values for \code{bw} and \code{n} are used if not provided in \code{density_args}.
Intersection density function
is then calculated by taking the integral of the minimum of the two functions,
from which the overlap outputs are calculated.
}
\examples{
#overlap of miles per gallon between 4-cylinder and 6-cylinder cars
library(dplyr)
a <- mtcars \%>\%
  filter(cyl==6)
a <- a[,1]
b <- mtcars \%>\%
  filter(cyl==4)
b <- b[,1]
pairwise_overlap(a,b)

}
\references{
http://stats.stackexchange.com/questions/97596/how-to-calculate-overlap-between-empirical-probability-densities
}
\seealso{
\code{\link{circular_overlap}} to calculate circular overlap.
}
