% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_overlap.R
\name{pairwise_overlap}
\alias{pairwise_overlap}
\title{Overlap between Two Empirical Density Estimates}
\usage{
pairwise_overlap(a, b, norm = TRUE, bw = NULL, n = NULL)
}
\arguments{
\item{a}{a matrix dataset with nrows= n individuals, with a column containing 
one measurement of a certain trait.}

\item{b}{another matrix dataset with the same trait measurements to be compared 
against a.}

\item{norm}{If TRUE, assume data are normally distributed; if FALSE,
additional normalization step is carried out by multiplying each density 
entry by the length of each vector.}

\item{bw}{the smoothing bandwidth to be used. The kernels are scaled such
that this is the standard deviation of the smoothing kernel.}

\item{n}{the number of equally spaced points at which the density is to be
estimated.}
}
\value{
The funtion returns a vector of three values: 
\item{overlap}{the overlap area *2 divided by the sum of areas under the two 
functions.} 
\item{overlap_a}{the overlap area divided by area under the function generated from 
a.}
\item{overlap_b}{the overlap area diveided by area under the function generated 
from b.}
}
\description{
This function generates kernel density estimates for two datasets on a common
grid to calculate the area of overlap between the two estimates.
See \url{http://stats.stackexchange.com/questions/97596/how-to-calculate-overlap-between-empirical-probability-densities},
this function is derived from code posted in the answer by user mmk.
}
\details{
This function generates kernel density estimates using the density() 
function for two datasets on a common grid.If bw = NULL, the default 'nrd0'is 
used. If n = NULL, the default value of 512 is used.Intersection density function 
is then calculated by taking the integral of the minimum of the two functions, 
from which the coefficients are calculated.
}
\examples{
#overlap of miles per gallon between 4-cylinder and 6-cylinder cars
library(ggplot2)
library(dplyr)
a <- mtcars \%>\%
  filter(cyl==6)
a <- a[,1]
b <- mtcars \%>\%
  filter(cyl==4)
b <- b[,1]
pairwise_overlap(a,b) 

}
\references{
http://stats.stackexchange.com/questions/97596/how-to-calculate-overlap-between-empirical-probability-densities
}
