% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitcher_traits.R
\docType{data}
\name{pitcher_traits}
\alias{pitcher_traits}
\title{Sample Multivariate Trait Data (Pitcher Plants)}
\format{
An object of class \code{data.frame} with 785 rows and 9 columns.
}
\usage{
pitcher_traits
}
\description{
Insert documentation here. Lines should be no more than 80 characters long.
}
\references{
Insert reference here, if there is a publication. Otherwise delete this line.
}
\keyword{datasets}
