% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_pool.R
\docType{data}
\name{reg_pool}
\alias{reg_pool}
\title{Sample Data for Ostats_regional}
\format{
An object of class \code{data.frame} with 7905 rows and 6 columns.
}
\usage{
reg_pool
}
\description{
This is a sample dataset with small mammal data from the NEON sites Bartlett
(BART) and Harvard Forest (HARV) for all available years.
The two sites belong to the same domain, the Northeast domain.
Each row represents an individual small mammal that was
trapped and weighed, with its species identity and body mass in grams.
}
\references{
Read, Q.D., J.M. Grady, P.L. Zarnetske, S. Record, B. Baiser, J. Belmaker, M.-N. Tuanmu, A. Strecker, L. Beaudrot, and K.M. Thibault. 2018. Among-species overlap in rodent body size distributions predicts species richness along a temperature gradient. *Ecography* 41(10):1718-1727. https://doi.org/10.1111/ecog.03641
}
\keyword{datasets}
