% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_mammal_Ostats.R
\docType{data}
\name{small_mammal_Ostats}
\alias{small_mammal_Ostats}
\title{Pre-calculated Overlap Statistics for NEON Small Mammals 2015}
\format{
An object of class \code{list} of length 4.
}
\usage{
small_mammal_Ostats
}
\description{
This is the pre-calculated output of the Ostats() function run on the
\code{\link{small_mammal_data}} dataset, after log-transforming the body masses.
}
\references{
Read, Q.D., J.M. Grady, P.L. Zarnetske, S. Record, B. Baiser, J. Belmaker, M.-N. Tuanmu, A. Strecker, L. Beaudrot, and K.M. Thibault. 2018. Among-species overlap in rodent body size distributions predicts species richness along a temperature gradient. *Ecography* 41(10):1718-1727. https://doi.org/10.1111/ecog.03641
}
\keyword{datasets}
